# imported/Appgard/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gard Security API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with RUT and access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /setup/admin:
    post:
      operationId: createInitialAdmin
      tags: [auth]
      summary: Create the first administrator when the database is empty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitialAdminBody"
      responses:
        "201":
          description: Initial administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Initial setup already completed
...
  /checkpoints:
    get:
      operationId: listCheckpoints
      tags: [checkpoints]
      summary: List checkpoints for a round
      parameters:
        - name: roundId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of checkpoints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkpoint"
    post:
      operationId: createCheckpoint
      tags: [checkpoints]
      summary: Create checkpoint (supervisor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckpointBody"
      responses:
        "201":
          description: Checkpoint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"
  /checkpoints/{id}/checkin:
    post:
      operationId: checkinCheckpoint
      tags: [checkpoints]
      summary: Check in at a checkpoint (guard proximity)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinBody"
      responses:
        "200":
          description: Checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkpoint"
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, investigating]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
      responses:
        "200":
          description: List of incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Report an incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentBody"
      responses:
        "201":
          description: Incident created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get incident details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
    patch:
      operationId: updateIncident
      tags: [incidents]
      summary: Update incident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentBody"
      responses:
        "200":
          description: Incident updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
...
  /location/update:
    post:
      operationId: updateLocation
      tags: [users]
      summary: Update guard GPS location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdateBody"
      responses:
        "200":
          description: Location updated
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        rut:
          type: string
        accessCode:
          type: string
      required:
        - rut
        - accessCode
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    InitialAdminBody:
      type: object
      properties:
        rut:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        accessCode:
          type: string
      required:
        - rut
        - name
        - email
        - phone
        - accessCode
    User:
      type: object
      properties:
        id:
          type: integer
        rut:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [guard, supervisor, admin]
        avatar:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rut
        - name
        - email
        - phone
        - role
        - active
        - createdAt
...
    Checkpoint:
      type: object
      properties:
        id:
          type: integer
        roundId:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        orderIndex:
          type: integer
        checkedIn:
          type: boolean
        checkedInAt:
          type: ["string", "null"]
          format: date-time
        checkedInById:
          type: ["integer", "null"]
      required:
        - id
        - roundId
        - name
        - latitude
        - longitude
        - radius
        - orderIndex
        - checkedIn
    CreateCheckpointBody:
      type: object
      properties:
        roundId:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        orderIndex:
          type: integer
      required:
        - roundId
        - name
        - latitude
        - longitude
        - radius
        - orderIndex
    CheckinBody:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude
    Incident:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, resolved]
        reportedById:
          type: integer
        reportedByName:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - title
        - description
        - priority
        - status
        - reportedById
        - reportedByName
        - createdAt
    CreateIncidentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        latitude:
          type: number
        longitude:
          type: number
      required:
        - title
        - description
        - priority
    UpdateIncidentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, resolved]
...
    LocationUpdateBody:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude
...