# imported/Appgard/lib/api-zod/src/generated/types/initialAdminBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gard Security API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InitialAdminBody {
  rut: string;
  name: string;
  email: string;
  phone: string;
  accessCode: string;
}